refactor: restore internal max query range resolution
