+++ src/proto/proto/cluster/search.proto
@@ -61,7 +61,6 @@ message SearchPartitionRequest {
     string          stream_type = 3;
     bytes               request = 5;
     bool   skip_max_query_range = 6;
-    int64      max_query_range = 7;
 }
 
 message SearchPartitionResponse {
