+++ src/search_service/src/streaming/cache.rs
@@ -230,7 +230,6 @@ pub async fn handle_cache_responses_and_deltas(
                     accumulated_results,
                     &mut curr_res_size,
                     user_id,
-                    max_query_range,
                     &mut remaining_query_range,
                     cached_search_duration,
                     cache_order_by,
@@ -280,7 +279,6 @@ pub async fn handle_cache_responses_and_deltas(
                 accumulated_results,
                 &mut curr_res_size,
                 user_id,
-                max_query_range,
                 &mut remaining_query_range,
                 cached_search_duration,
                 cache_order_by,
